requirements.txt
speedtest-cli
pytz